RFCT: separate functions to their respective concerns for maintainability
