+++ backend/src/apis/v2/helpers/image_process_utils.py
@@ -1,3 +1,4 @@
+import math
 import cv2
 import numpy as np
 from sqlalchemy.orm import Session
@@ -60,6 +61,14 @@ def process_chip(
         image_settings.crop_size,
     )
     return chip_processor
+def extract_hsv_mask_and_area_sum(image: np.ndarray) -> tuple[np.ndarray, float]:
+    """Extracts the HSV mask and calculates the area of all contours in the given image."""
+    hsv = cv2.cvtColor(image, cv2.COLOR_BGR2HSV_FULL)
+    hsv_mask = cv2.inRange(hsv, np.array([1, 0, 0]), np.array([254, 255, 255]))
+
+    area_sum = np.count_nonzero(hsv_mask)
+
+    return hsv_mask, area_sum
 
 
 def check_single(
@@ -82,3 +91,33 @@ def check_single(
             return ContourHandler.filter_and_build_contour_info(new_contours)
 
     return ContourList(contours=[contour_info])
+
+
+def create_focus_chip_mask(image: np.ndarray) -> np.ndarray:
+    """Creates a mask of the focus chip area by cropping the image and adding borders."""
+    focus_chip_pad = math.floor(image.shape[0] / 4 * 0.7)
+    cropped_image = image[
+        focus_chip_pad:-focus_chip_pad, focus_chip_pad:-focus_chip_pad
+    ]
+
+    _, border_gray, _, _ = create_border(cropped_image, padding=focus_chip_pad)
+
+    mask_handler = MaskHandler(border_gray)
+    return mask_handler.binary_image
+
+
+def get_largest_info_and_mask(mask_image: np.ndarray) -> tuple[ContourInfo, np.ndarray]:
+    """Finds and returns the largest contour's information and its corresponding mask."""
+    contour_info_list = create_contour_list(mask_image)
+    largest_contour_info = max(contour_info_list.contours, key=lambda x: x.area)
+
+    largest_mask = np.zeros(mask_image.shape[:2], np.uint8)
+    cv2.drawContours(
+        largest_mask,
+        [largest_contour_info.contour],
+        -1,
+        BGRColors.WHITE.value,
+        -1,
+    )
+
+    return largest_contour_info, largest_mask
