+++ backend/src/apis/v2/helpers/processor/defect_processor.py
@@ -1,13 +1,9 @@
 import cv2
 import numpy as np
 
-from apis.v2.components.extract_roi import (
-    get_defect_color,
-    create_focus_chip_mask,
-    determine_size,
-    get_hsv_mask_and_contour_area,
-    get_largest_info_and_mask,
-)
+from apis.v2.helpers.image_process_utils import create_focus_chip_mask, extract_hsv_mask_and_area_sum, get_largest_info_and_mask
+from constants.colors import CSAMcolor
+from constants.image_thresholds import AugmentThreshold
 from constants.tf_model import ClassLabel
 from constants.chip_thresholds import ChipThreshold
 from interface.image_process import ChipProcessorInterface
@@ -77,7 +73,7 @@ class DefectProcessor:
         major_binary_roi = cv2.bitwise_and(
             rotated_image, rotated_image, mask=major_binary_mask
         )
-        _, binary_area_sum = get_hsv_mask_and_contour_area(major_binary_roi)
+        _, binary_area_sum = extract_hsv_mask_and_area_sum(major_binary_roi)
 
         if binary_area_sum == 0:
             return ClassLabel.G.value, None, None
@@ -94,13 +90,54 @@ class DefectProcessor:
             rotated_image, rotated_image, mask=major_defect_mask
         )
 
-        defect_mask, defect_area_sum = get_hsv_mask_and_contour_area(major_defect_roi)
+        defect_mask, defect_area_sum = extract_hsv_mask_and_area_sum(major_defect_roi)
 
         defect_image = cv2.bitwise_and(rotated_image, rotated_image, mask=defect_mask)
-        defect_color = get_defect_color(defect_image)
+        defect_color = self._get_defect_color(defect_image)
 
-        defect_size = determine_size(defect_area_sum, major_defect_contour_info.area)
+        defect_size = self._determine_size(
+            defect_area_sum, major_defect_contour_info.area
+        )
 
         if defect_color is None or defect_size is None:
             return ClassLabel.OTHERS.value, None, None
         return ClassLabel.NG.value, defect_color, defect_size
+
+    def _get_defect_color(defect: np.ndarray) -> str | None:
+        """Determine the most common color type of a defect, excluding black (background)."""
+
+        color_hexes, hex_count = np.unique(
+            defect.reshape(-1, defect.shape[-1]), axis=0, return_counts=True
+        )
+
+        # Remove black (background) color
+        color_hexes = np.delete(color_hexes, 0, axis=0)
+        hex_count = np.delete(hex_count, 0, axis=0)
+
+        if color_hexes.size == 0:
+            return None
+
+        # Find the most common color in the defect
+        most_common_color = color_hexes[np.argmax(hex_count)]
+
+        # Match the high count color to the csam_color dictionary adn return color type
+        return next(
+            (
+                csam_color.value.name
+                for csam_color in CSAMcolor
+                if np.array_equal(csam_color.value.bgr, most_common_color)
+            ),
+            None,
+        )
+
+    def _determine_size(area_sum: float, defect_area: float) -> str | None:
+        """Determines the size classification based on the area ratio."""
+        size_ratio = round(area_sum / defect_area * 100, 2)
+
+        if size_ratio == 0:
+            return None
+        elif size_ratio < AugmentThreshold.SMALL_SIZE_THRESHOLD.value:
+            return "small"
+        elif size_ratio > AugmentThreshold.BIG_SIZE_THRESHOLD.value:
+            return "big"
+        return "medium"
